# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Neon Murer CMS API
  description: |
    Vollständige REST API für das Neon Murer Content Management System.
    
    ## Authentifizierung
    Die meisten Endpoints erfordern JWT-Token Authentication über den `Authorization: Bearer <token>` Header.
    
    ## Fehlerbehandlung
    Alle Endpoints verwenden einheitliche HTTP-Status-Codes und JSON-Antworten.
    
    ## Rate Limiting
    API-Aufrufe sind auf 100 Requests pro Minute pro IP-Adresse begrenzt.
    
  version: 1.0.0
  contact:
    name: Neon Murer AG
    url: https://neonmurer.ch
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api
    description: Development Server
  - url: https://cms.neonmurer.ch/api  
    description: Production Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MediaFile:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
          example: "image-1738255046223.jpg"
        originalName:
          type: string
          example: "original-image.jpg"
        path:
          type: string
          example: "content/images/image-1738255046223.jpg"
        size:
          type: integer
          example: 256000
        mimeType:
          type: string
          example: "image/jpeg"
        extension:
          type: string
          example: "jpg"
        altText:
          type: string
        description:
          type: string
        optimized:
          type: boolean
          example: true
        compressionRatio:
          type: number
          example: 0.75
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /media/upload:
    post:
      tags: [Media]
      summary: Dateien hochladen
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Upload erfolgreich
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  files:
                    type: array
                    items:
                      $ref: '#/components/schemas/MediaFile'
                  message:
                    type: string
                  totalSavings:
                    type: string
...
  /media/{filename}:
    delete:
      tags: [Media]
      summary: Datei löschen
      security:
        - bearerAuth: []
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Datei erfolgreich gelöscht
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    
  # SETTINGS ENDPOINTS
...